+++ agentica/tools/url_crawler_tool.py
@@ -1,7 +1,9 @@
 # -*- coding: utf-8 -*-
 """
-@author:XuMing(xuming624@qq.com)
-@description: URL Crawler Tool
+URL Crawler Tool.
+
+Requires the [crawl] extras:
+    pip install agentica[crawl]
 """
 import asyncio
 import hashlib
@@ -11,9 +13,14 @@ import json
 from urllib.parse import ParseResult, urlparse, urljoin
 
 import httpx
-from bs4 import BeautifulSoup
-import sys
-sys.path.append("../..")
+try:
+    from bs4 import BeautifulSoup
+except ImportError as e:
+    raise ImportError(
+        "UrlCrawlerTool requires the [crawl] extras. Install with:\n\n"
+        "    pip install agentica[crawl]\n"
+    ) from e
+
 from agentica.tools.base import Tool
 from agentica.config import AGENTICA_HOME
 from agentica.utils.log import logger